+++ src/azure-cli/HISTORY.rst
@@ -36,6 +36,10 @@ Release History
 * Auto-completers can now select resources from any resource group. It can be a different resource group than the one specified with "-g"
 * Added support for "--sub-domain-suffix" and "--disable_gateway_auth" parameters in the "az hdinsight application create" command
 
+**Managed Services**
+
+* Introducing managed service command module in preview.
+
 **Profile**
 * Suppress `--subscription` argument for logout command.